x11: fix ICC profiling for multiple monitors
To find the correct ICC profile, the screen number was calculated
directly from the xrandr order of the screens.
But if a primary screen is set, it should be the first xinerama screen,
even if it is not the first xrandr screen.

Sorting the displays to put the primary screen as the first one.